f2fs: rename /sys/fs/f2fs
fsmgr and vold tries to read nodes from here and run GC by itself.

Since it conflicts with rapid GC(not to mention less effective),
rename /sys/fs/f2fs to /sys/fs/f2fs_dev to avoid Android reading it entirely.

We could rename gc_urgent and dirty_segments but this approach seems better
as it'll still be detected as a f2fs volume and try to run GC.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Change-Id: Ib2f0c81b8a80f340f0bdecdf35dd3cf73e0b500c